随笔管理前后端完成

+++ ruoyi-modules/biz-blog/src/main/java/com/ruoyi/blog/domain/Blog.java
@@ -32,9 +32,9 @@ public class Blog extends BaseEntity {
 
     private Long commentCnt;
 
-    private Integer articleClassify;
+    private Long articleClassify;
 
-    private Integer personClassify;
+    private Long personClassify;
 
     private Integer type;
 
@@ -118,19 +118,19 @@ public class Blog extends BaseEntity {
         this.title = title;
     }
 
-    public Integer getArticleClassify() {
+    public Long getArticleClassify() {
         return articleClassify;
     }
 
-    public void setArticleClassify(Integer articleClassify) {
+    public void setArticleClassify(Long articleClassify) {
         this.articleClassify = articleClassify;
     }
 
-    public Integer getPersonClassify() {
+    public Long getPersonClassify() {
         return personClassify;
     }
 
-    public void setPersonClassify(Integer personClassify) {
+    public void setPersonClassify(Long personClassify) {
         this.personClassify = personClassify;
     }
